requirements.txt
pygame-ce==2.5.3
jupyter~=1.1.1
pip~=25.0.1
attrs~=25.1.0
wheel~=0.45.1
pillow~=11.1.0
docutils~=0.21.2
filelock~=3.17.0
numpy~=2.2.1
pyinstaller~=6.12.0